refactor(catalog): rename CatalogWidgetStore to CatalogDisplayStore
