+++ src/stores/Widgets/CatalogWidget/CatalogDisplayStore.ts
