+++ src/stores/Widgets/index.ts
@@ -1,6 +1,6 @@
 export * from "./AnimatorWidgetStore/AnimatorWidgetStore";
+export * from "./CatalogWidget/CatalogDisplayStore";
 export * from "./CatalogWidget/CatalogPlotWidgetStore";
-export * from "./CatalogWidget/CatalogWidgetStore";
 export * from "./EmptyWidgetStore";
 export * from "./HistogramWidgetStore/HistogramWidgetStore";
 export * from "./LayerListWidgetStore/LayerListWidgetStore";
